refactor features hooks to easily reuse
