+++ applications/fishing-map/src/features/map/map-viewport.hooks.ts
@@ -40,20 +40,16 @@ const viewportState = atom<MapCoordinates>({
 export default function useViewport(): UseViewport {
   const [viewport, setViewport] = useRecoilState(viewportState)
 
-  const setMapCoordinates = useCallback(
-    (viewport: SetMapCoordinatesArgs) => {
-      setViewport({ ...viewport })
-    },
-    [setViewport]
-  )
+  const setMapCoordinates = useCallback((viewport: SetMapCoordinatesArgs) => {
+    setViewport({ ...viewport })
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [])
 
-  const onViewportChange = useCallback(
-    (viewport: ViewportProps) => {
-      const { latitude, longitude, zoom } = viewport
-      setViewport({ latitude, longitude, zoom })
-    },
-    [setViewport]
-  )
+  const onViewportChange = useCallback((viewport: ViewportProps) => {
+    const { latitude, longitude, zoom } = viewport
+    setViewport({ latitude, longitude, zoom })
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [])
 
   return { viewport, onViewportChange, setMapCoordinates }
 }
