+++ applications/fishing-map/src/features/map/map.hooks.ts
@@ -1,5 +1,5 @@
 import { useSelector, useDispatch } from 'react-redux'
-import { useEffect, useMemo, useRef, useState } from 'react'
+import { useRef } from 'react'
 import { ExtendedFeatureVessel, InteractionEvent } from '@globalfishingwatch/react-hooks'
 import { Generators, TimeChunks } from '@globalfishingwatch/layer-composer'
 import { ContextLayerType, Type } from '@globalfishingwatch/layer-composer/dist/generators/types'
@@ -230,95 +230,4 @@ export const useCurrentTimeChunkId = () => {
   const currentTimeChunks = style?.metadata?.temporalgrid?.timeChunks as TimeChunks
   const currentTimeChunkId = currentTimeChunks?.activeId
   return currentTimeChunkId
-}
-
-export const useMapSourceLoaded = (sourceId: string, cacheKey?: string) => {
-  const mapInstance = useMapboxInstance()
-  const [sourceLoaded, setSourceLoaded] = useState(false)
-
-  useEffect(() => {
-    if (cacheKey) {
-      setSourceLoaded(false)
-    }
-  }, [cacheKey])
-
-  useEffect(() => {
-    const sourceLoadingCallback = (sourcedataEvent: any) => {
-      if (sourcedataEvent.sourceId === sourceId) {
-        setSourceLoaded(false)
-      }
-    }
-
-    if (!cacheKey) {
-      if (mapInstance && sourceLoaded) {
-        mapInstance.on('sourcedataloading', sourceLoadingCallback)
-      }
-    }
-
-    return () => {
-      if (mapInstance) {
-        mapInstance.off('sourcedataloading', sourceLoadingCallback)
-      }
-    }
-  }, [cacheKey, mapInstance, sourceId, sourceLoaded])
-
-  useEffect(() => {
-    const sourceCallback = () => {
-      if (mapInstance?.getSource(sourceId) && mapInstance?.isSourceLoaded(sourceId)) {
-        setSourceLoaded(true)
-        mapInstance.off('idle', sourceCallback)
-      }
-    }
-
-    const sourceErrorCallback = (errorEvent: any) => {
-      if (
-        errorEvent.sourceId === sourceId &&
-        mapInstance?.getSource(sourceId) &&
-        mapInstance?.isSourceLoaded(sourceId)
-      ) {
-        setSourceLoaded(true)
-      }
-    }
-
-    if (mapInstance && sourceId && !sourceLoaded) {
-      mapInstance.on('idle', sourceCallback)
-      mapInstance.on('error', sourceErrorCallback)
-    }
-
-    return () => {
-      if (mapInstance) {
-        mapInstance.off('idle', sourceCallback)
-        mapInstance.off('error', sourceErrorCallback)
-      }
-    }
-  }, [mapInstance, sourceId, sourceLoaded])
-
-  return sourceLoaded
-}
-
-export const useMapFeatures = ({
-  sourceId,
-  sourceLayer = 'main',
-  cacheKey,
-  filter,
-}: {
-  sourceId: string
-  sourceLayer?: string
-  cacheKey?: string
-  filter?: any[]
-}) => {
-  const mapInstance = useMapboxInstance()
-  const sourceLoaded = useMapSourceLoaded(sourceId, cacheKey)
-
-  const features = useMemo(() => {
-    if (sourceLoaded) {
-      const features = mapInstance?.querySourceFeatures(sourceId, {
-        sourceLayer: sourceLayer,
-        ...(filter && { filter }),
-      })
-      return features
-    }
-  }, [sourceLoaded, mapInstance, sourceId, sourceLayer, filter])
-
-  return { features, sourceLoaded }
 }
